Move the play history model into the spotify package
HistoryEntry is TrackRef's counterpart -- a record of what Spotify said about a
track -- and TrackRef lives in the spotify package, so the two siblings had no
business being in different ones. What a context uri looks like, and which
contexts accept an offset, is knowledge about Spotify rather than about fzf.

The entry now holds the track it was given rather than a rendered row.
Composing the row and then splitting it again to recover the track's name is
the re-parsing the candidate line format was reworked to remove; deciding what
reaches a row belongs to the renderer, which is where it now happens -- and a
playlist name is cleaned there too, so a name with a newline in it can no
longer break the action menu the way it could not break the row.

screens.py loses 83 lines and the interface package stops carrying a model.
